[move-compiler] Move constant copy resolution and dependency walking into dedicated homes
Context::constant_copy now owns the full resolve-or-synthesize logic
(registry hit, copy synthesis, unfoldable/outside-compilation errors),
so the CFGIR rewrite pass devolves into rewrite-or-UnresolvedError with
an ICE assert. dependent_constants moves to cfgir/constants.rs as
compute_dependent_constants alongside the other constant walkers.
